Add a "Destructive Calls" flag to make it harder to do the wrong thing (e.g. remove a repo)

+++ github/resource_repository.go
@@ -19,6 +19,7 @@ package github
 import (
 	"context"
 	"errors"
+	"fmt"
 	"reflect"
 
 	gitprovider "github.com/fluxcd/go-git-providers"
@@ -134,6 +135,11 @@ func (r *userRepository) Reconcile(ctx context.Context) (bool, error) {
 //
 // ErrNotFound is returned if the resource doesn't exist anymore.
 func (r *userRepository) Delete(ctx context.Context) error {
+	// Don't allow deleting repositories if the user didn't explicitely allow dangerous API calls.
+	if !r.destructiveActions {
+		return fmt.Errorf("cannot delete repository: %w", ErrDestructiveCallDisallowed)
+	}
+
 	_, err := r.c.Repositories.Delete(ctx, r.ref.GetIdentity(), r.ref.GetRepository())
 	return handleHTTPError(err)
 }
